# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 法律合同RAG切分工具
  description: 用于将法律合同文档智能切分为适合检索增强生成（RAG）应用的文本块
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
  /api/v1/chunk:
    post:
      tags:
        - 文档切分
      summary: 上传合同文件并获取切分结果
      operationId: chunk_file
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChunkRequest'
        required: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkResultResponse'
        '415':
          description: 不支持的文件格式
  /api/v1/health:
    get:
      tags:
        - 系统
      summary: 健康检查
      operationId: health_check
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/v1/info:
    get:
      tags:
        - 系统
      summary: 获取服务信息
      operationId: get_info
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
components:
  schemas:
    ChunkRequest:
      type: object
      required:
        - file
        - request_id
      properties:
        file:
          type: string
          format: binary
          description: 合同文件（PDF/DOCX/TXT）
        request_id:
          type: string
          description: 请求唯一标识
        max_tokens:
          type: integer
          default: 500
          description: 每个chunk最大token数
        overlap_tokens:
          type: integer
          default: 50
          description: 相邻chunk重叠token数
        min_tokens:
          type: integer
          default: 100
          description: 最小token数
    ChunkMetadata:
      type: object
      properties:
        title:
          type: string
          description: 章节标题
        type:
          type: string
          description: 类型
        level:
          type: integer
          description: 标题级别
        page_number:
          type: integer
          description: 页码
        file_name:
          type: string
          description: 文件名
    ChunkResponse:
      type: object
      properties:
        request_id:
          type: string
          description: 请求标识
        chunk_index:
          type: integer
          description: chunk索引
        content:
          type: string
          description: chunk文本内容
        metadata:
          $ref: '#/components/schemas/ChunkMetadata'
        token_count:
          type: integer
          description: token数
        previous_chunk_index:
          type: integer
          nullable: true
          description: 上一个chunk索引
        next_chunk_index:
          type: integer
          nullable: true
          description: 下一个chunk索引
    ChunkResultData:
      type: object
      properties:
        request_id:
          type: string
        file_name:
          type: string
        total_chunks:
          type: integer
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkResponse'
    ChunkResultResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/ChunkResultData'
    HealthResponse:
      type: object
      properties:
        status:
          type: string
        version:
          type: string
    InfoResponse:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        supported_formats:
          type: array
          items:
            type: string
        parameters:
          type: object